feat: refactor database user environment variable and enhance middleware for request ID handling

+++ src/common/infra/database/seed/clean.ts
@@ -4,9 +4,9 @@ import { sql } from "drizzle-orm/sql/sql";
 import fs from "fs";
 import path from "path";
 
-const envFile = `.env${process.env.NODE_ENV ? `.${process.env.NODE_ENV}.local` : ""}`;
+const envFile = `.env${process.env.NODE_ENV ? `.${process.env.NODE_ENV}` : ""}`;
 dotenv.config({ path: envFile });
-const connString = `postgresql://${process.env.DATABASE_USERNAME}:${process.env.DATABASE_PASSWORD}@${process.env.DATABASE_HOST}:${process.env.DATABASE_PORT}/${process.env.DATABASE_NAME}`;
+const connString = `postgresql://${process.env.DATABASE_USERNAMENAME}:${process.env.DATABASE_PASSWORD}@${process.env.DATABASE_HOST}:${process.env.DATABASE_PORT}/${process.env.DATABASE_NAME}`;
 const db = drizzle(connString);
 const migrationsDir = path.resolve(__dirname, "../migrations");
 
@@ -14,13 +14,17 @@ async function cleanDatabase() {
 	console.log("<--- Resetting the database");
 	try {
 		// delete only all folder in the migrations directory, not the directory itself
-		const files = fs.readdirSync(migrationsDir);
-		for (const file of files) {
-			const filePath = path.join(migrationsDir, file);
-			if (fs.lstatSync(filePath).isDirectory()) {
-				fs.rmSync(filePath, { recursive: true, force: true });
-				console.log(" -- Deleted migration directory:", filePath);
+		if (fs.existsSync(migrationsDir)) {
+			const files = fs.readdirSync(migrationsDir);
+			for (const file of files) {
+				const filePath = path.join(migrationsDir, file);
+				if (fs.lstatSync(filePath).isDirectory()) {
+					fs.rmSync(filePath, { recursive: true, force: true });
+					console.log(" -- Deleted migration directory:", filePath);
+				}
 			}
+		} else {
+			console.log(" -- Migrations directory does not exist, skipping migration cleanup:", migrationsDir);
 		}
 
 		// deactivate the referential integrity to avoid issues with foreign keys
